Fix build without glslang
While building Anvil without GLSLANG, we get a compilation error.
This commit fixes this compilation error by not defining the
get_disassembly function in the ShaderModule class, which
was trying to return the m_disassembly variable (which is also
not defined in this case). Also the commit removes depencies on
the GLSLANG headers in this case.